requirements.txt

Django==3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==3.13.0
django-extensions==3.1.5


pymongo==4.13.0
dnspython==2.7.0
mongoengine==0.29.1
python-dotenv==1.1.1
certifi==2025.6.15


numpy==1.23.5
pandas==2.0.
scikit-learn==1.2.2
xgboost==1.7.6

tensorflow-cpu==2.12.0
h5py==3.8.0


gunicorn==23.0.0
